+++ mcserver/constants.py
@@ -3,5 +3,5 @@ __version__: str = "2.0.0"
 modrinth_base_api: str = "https://api.modrinth.com/v2"
 papermc_base_api: str = "https://fill.papermc.io"
 
-si_prefixes = ["", "K", "M", "B", "T"]
-iec_prefixes = ["B", "KiB", "MiB", "GiB", "TiB"]
+si_prefixes: list = ["", "K", "M", "B", "T"]
+iec_prefixes: list = ["B", "KiB", "MiB", "GiB", "TiB"]