+++ mcserver/metadata.py
@@ -1,0 +1,13 @@
+#Modules
+# Standard
+import json
+
+from pathlib import Path
+
+# MCServer
+from shared import mcserver_dir
+
+#Paths
+metadata_file = mcserver_dir / "metadata.json"
+
+#Functions