+++ src/mcserver/commands/server_status.py
@@ -2,5 +2,6 @@ import argparse
 
 
 def main(args: argparse.Namespace) -> int:
-    print(args)
-    return 0
+    import logging as log
+    log.error("Command is not implemented yet.")
+    return 1