+++ src/mcserver/commands/whitelist_add.py
@@ -3,5 +3,6 @@ import argparse
 
 def main(args: argparse.Namespace) -> int:
     players: list[str] = args.players
-    print(args)
-    return 0
+    import logging as log
+    log.error("Command is not implemented yet.")
+    return 1