+++ src/mcserver/commands/whitelist_remove.py
@@ -1,31 +1,32 @@
 import argparse
-from uuid import UUID
-from ..server_files.whitelist import remove_players
-from ..minecraft.player_identity import PlayerID
-
-import logging as log
-
 
 def main(args: argparse.Namespace) -> int:
     players: list[str] = args.players
-    player_ids: list[PlayerID] = []
-    for player in players:
-        try:
-            UUID(player)
-            log.debug(f"{player} is likely an UUID")
-            player_id: PlayerID = {"player_uuid": player}
+    import logging as log
 
-            player_ids.append(player_id)
-        except ValueError:
-            log.debug(f"{player} is likely a name")
-            player_id: PlayerID = {"player_name": player}
+    # from uuid import UUID
+    # from ..server_files.whitelist import remove_players
+    # from ..minecraft.player_identity import PlayerID
 
-            player_ids.append(player_id)
+    log.error("Command is not implemented yet.")
+    # player_ids: list[PlayerID] = []
+    # for player in players:
+    #     try:
+    #         UUID(player)
+    #         log.debug(f"{player} is likely an UUID")
+    #         player_id: PlayerID = {"player_uuid": player}
 
-    try:
-        remove_players(player_ids)
-    except FileNotFoundError as error:
-        log.error(error)
-        return 1
+    #         player_ids.append(player_id)
+    #     except ValueError:
+    #         log.debug(f"{player} is likely a name")
+    #         player_id: PlayerID = {"player_name": player}
 
-    return 0
+    #         player_ids.append(player_id)
+
+    # try:
+    #     remove_players(player_ids)
+    # except FileNotFoundError as error:
+    #     log.error(error)
+    #     return 1
+
+    return 1