+++ CHANGELOG.md
@@ -7,11 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.0] - 2026-08-18
+
 ### Added
 
-- Paper and Purpur as server software
+- Command to stop the Minecraft server
+- Paper and Purpur server software
 - Metadata to keep information of what version of MCServer is used
-- State to keep information of a running MCServer to avoid conflict
+- State to keep information of a running MCServer to avoid conflicting command
+- Listing whitelisted players
+- Listing operator players
+
+### Changed
+
+- Uh... add, search, and show, are being worked on (they're not available for now)
+- The codebase is now bigger with small gain (for the user)
 
 ## [2.0.0] - 2026-06-14